Big change. There are now 4 components:
- front-main: The main frontend application
- front-login: just a login page
- back/tristram: main backend api
- back/auth: backend authentication

The idea is that the main components now have nothing to do with authentication. If running this in your home, you've no need for authentication so that simplifies a lot. If you want authentication, that's what the other two pieces are for.  But access control is mainly handled by nginx -- the /app/ and /api/ routes are all protected by auth (if using that image).

Along with that, everything is now in one container (whether or not that's recommended). The container runs 2 or 3 processes. nginx is one, and there either 1 or 2 gunicorns running depending on whether authentication is being used. Being in the same container makes it one tidy little package.